🦺 Add safer access to settings key
The Wordle settings key prefix changed again, which was causing the
extension to stop working altogether due to an unsafe access. This
change makes it so that the extension should continue working if the
settings key changes again (though settings preferences such as dark
mode and high contrast mode might stop working in that case).

Fixes #8